+++ include/XML/XMLSerializer.hpp
@@ -153,12 +153,12 @@ namespace claid
                 this->callInt(property, m);
             }
 
-            void count(const std::string& name, size_t& count)
+            void count(const std::string& name, int32_t& count)
             {
                 // Do nothing
             }
 
-            void countElements(size_t& count)
+            void countElements(int32_t& count)
             {
                 // Do nothing
             }
